+++ intents/__init__.py
@@ -5,6 +5,7 @@ from typing import List
 __version__ = "0.4.dev1"
 
 from intents.language_codes import LanguageCode
+from intents.geolocation import Location
 
 from intents.model.intent import Intent, FulfillmentContext, FulfillmentResult
 from intents.model.entity import Entity, EntityMixin, Sys